requirements.txt
awslambdaric==2.0.4
requests==2.28.2
Pillow==9.4.0
